Extract Video Control
